refatoração do startup da aplicação mudando o program.cs para ficar mais limpo
