+++ WebAPI/Program.cs
@@ -1,24 +1,16 @@
 using Microsoft.AspNetCore.Mvc;
 using WebAPI;
+using WebAPI.Startup;
 
 var builder = WebApplication.CreateBuilder(args);
-builder.Services.AddSingleton<IProdutosService, ProdutosService>();
+
+builder.Services.ConfigureDI();
 
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 var app = builder.Build();
 
-app.MapGet("produtos", async  ( [FromServices] IProdutosService _service) =>
-{
-    var lista = _service.GetAll();
-    return Results.Ok(lista);
-});
-
-app.MapGet("produtos/{id}", async ([FromServices] IProdutosService _service, [FromRoute] int id) =>
-{
-    var item = _service.GetOne(id);
-    return Results.Ok(item);
-});
+app.AddEndpoints();
 
 app.UseSwagger();
 app.UseSwaggerUI();
