+++ WebAPI/Startup/DependencyInjectionConfig.cs
@@ -1,0 +1,11 @@
+﻿namespace WebAPI.Startup
+{
+    public static class DependencyInjectionConfig
+    {
+        public static void ConfigureDI(this IServiceCollection services)
+        {
+            services.AddSingleton<IProdutosService, ProdutosService>();
+            // poderemos colocar aqui todos os serviços e repositórios
+        }
+    }
+}
